// src/models/models.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Time: any;
  Upload: any;
};
// ...
export type EyewearShop = {
  __typename?: "EyewearShop";
  id: Scalars["ID"];
  title: Scalars["String"];
  address: Scalars["String"];
  region: Scalars["String"];
  country: Scalars["String"];
  phone: Scalars["String"];
  inHouse: Scalars["Boolean"];
  active: Scalars["Boolean"];
};

export type EyewearShopConnection = Connection & {
  __typename?: "EyewearShopConnection";
  totalCount: Scalars["Int"];
  pageInfo: PageInfo;
  edges: Array<Maybe<EyewearShopEdge>>;
};

export type EyewearShopEdge = {
  __typename?: "EyewearShopEdge";
  node: EyewearShop;
};
// ...
export type PaginationInput = {
  page: Scalars["Int"];
  size: Scalars["Int"];
};
// ...
export type QueryEyewearShopsArgs = {
  page: PaginationInput;
};